+++ playground/src/main/java/com/chenbitao/word/playground/demo/visio/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Visio drawing read demos.
+ */
+package com.chenbitao.word.playground.demo.visio;