+++ word-generator/src/main/java/com/chenbitao/word/exception/VisioDrawingException.java
@@ -1,0 +1,11 @@
+package com.chenbitao.word.exception;
+
+/**
+ * Visio drawing read exception.
+ */
+public class VisioDrawingException extends RuntimeException {
+
+    public VisioDrawingException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}